third_party/WebKit/LayoutTests/http/tests/asmjs/asm-warnings-expected.txt

CONSOLE WARNING: line 13: Invalid asm.js: Bad variable declaration
CONSOLE WARNING: line 5: Invalid asm.js: Can only use immutable variables in global definition
